refactor: extract DifficultyLevel and defaultProgressState to shared modules
Move DifficultyLevel enum to types/index.ts and defaultProgressState to
state/constants.ts so they can be imported by e2e tests without pulling
in Jotai atoms and other app-specific dependencies.
